requirements.txt
flask
werkzeug
pillow
python-magic